requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
asyncio
attrs
confluent-kafka
dnspython
frozenlist
idna
logging
multidict
propcache
psycopg2-binary
py4j
pymongo
pyspark
spark
typing
yarl
apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-postgres
requests